### 2019-11-23 #11 @mcbaneg
10 year reproducibility challenge: reproduction of J. Chem. Phys. 112, 4417 (2000)
[article.pdf](https://github.com/ReScience/submissions/files/3882888/article.pdf)

**Original article:** [Gottfried and McBane, J. Chem. Phys. 112, 4417 (2000)](https://scholarworks.gvsu.edu/chm_articles/13/)

**Code URL:**  https://github.com/mcbaneg/tenyears-submission (also: http://faculty.gvsu.edu/mcbaneg/virial6-resciencec.zip)

**Scientific domain:** Chemical Physics
**Programming language:** Fortran
