### 2019-11-25 #11 @mcbaneg
Okay, I have modifed the PDF link to point to the .pdf of the draft that I attached to the posting (that is, uploaded to GitHub.)
I was unable to do the same with the metadata.yaml file because when I try to attach it I get a message "We don't support that file type,  Try again with [list of file extensions it accepts, including .zip]"  So, I made a .zip file out of metadata.yaml and uploaded that; I hope that works.

The original metadata.yaml is present at the site I originally listed, but either the web server doesn't automatically start a download when a link to a .yaml file is followed, or the browser doesn't recognize that as what it should do.

If there's a way to upload the metadata.yaml file to the GitHub submission, I'd be happy to do it to save the user the hassle of unzipping.

-G.
