### 2019-11-26 #11 @khinsen
Good question. I have never used the GitHub file upload feature for anything but images. Most ReScience authors create a GitHub repository for their submission, which contains all their code plus the article source code and the metadata. That makes it easy to put links to any file anywhere, and also allows reviewers and editors to propose modifications (which, as an editor, I do for the metadata in the end of the publication process). Another advantage is that GitHub repositories are automatically archived by [Software Heritage](https://www.softwareheritage.org/), so they are safe in case of GitHub shutting down, being hackes, changing their rules, etc.

Anyway, the ZIP you uploaded for the metadata works well for me!
