### 2019-11-26 #11 @khinsen
@pdebuyl Would you willing to review this paper?

Note that it is a special case: the first submission to the Ten Years Challenge, written in the absence of any specific instructions from ReScience. With the consentment of the author, we plan to use this article and the review to prepare more detailed guidelines for other authors.

Also, since we expect to have many submissions to this special issue, we will try a revised reviewing procedure as well, with a single reviewer.
