### 2019-11-27 #11 @pdebuyl
Hi @khinsen I will review. I'll take notes along the way :-)
