### 2019-11-27 #11 @mcbaneg
Pierre,

I do know git, though I have not used GitHub much.  I’ll set up a repository.

Do you have a BLAS library handy to use, or would it help for me to put the BLAS reference sources there for you?

-George.

George C. McBane (mcbaneg@gvsu.edu<mailto:mcbaneg@gvsu.edu>)
Professor of Chemistry
Assistant Dean for Research, Facilities, and Analytics
College of Liberal Arts and Sciences
Grand Valley State University

B-4-229 Mackinac Hall, (616) 331-2506
Chemistry site http://www.gvsu.edu/chem/
CLAS site http://www.gvsu.edu/clas/
individual site http://faculty.gvsu.edu/mcbaneg/
