+++ angular.f
@@ -1,0 +1,225 @@
+c  this file holds functions stolen from Molscat that are
+c  needed for evaluating 4D angular quantum corrections
+
+      FUNCTION YRR(L1,L2,L,CT1,CT2,DP)
+C
+C     BISPHERICAL HARMONIC ANGULAR FUNCTIONS FOR TWO DIATOMS
+C     CT1, CT2 ARE COS(THETA-1) AND COS(THETA-2), AND
+C          DP IS DELTA(PHI), I.E., PHI2-PHI1, IN RADIANS
+C     CF. GREEN, JCP 62, 2271 (1975) APPENDIX.
+C     N.B. P(L,M;X) THERE IS (2*PI)**-1/2 NORMALIZED P(L,M;X)
+C          MOLSCAT PLM(L,M,CT) ROUTINE IS NORMALIZED ON CT, AND
+C               PLM(L,0,1.D0)=SQRT((2L+1)/2) .
+C          THUS, MUST MULT EACH PLM BY (2*PI)**-1/2
+C
+C     ODD  L1+L2+L  *NOT* ALLOWED; TRAPPED W/MESSAGE AND STOP
+C
+C     NEEDS ROUTINES THRJ(XJ1,XJ2,XJ3,XM1,XM2,XM3)
+C                    PLM(L,M,COSTH)
+C                    PARITY(J)
+C
+      IMPLICIT DOUBLE PRECISION (A-H,O-Z)
+      LOGICAL ODD
+! added 13 Nov 2019 to avoid collision with new reserved name
+      external parity
+      DATA PI/3.14159 26535 89793 D0/
+      ODD(I)=2*(I/2)-I.NE.0
+C
+      IF (ODD(L1+L2+L)) GO TO 9999
+C
+      XL1=L1
+      XL2=L2
+      XL=L
+C     SQRT(4*PI) FROM Y(L,M,THETA=0), 2*PI FOR TWO PLM'S
+      DEN=SQRT(4.D0*PI)*2.D0*PI
+      FACT=((2.D0*XL+1.D0)/DEN)*PARITY(L1+L2)
+      MTOP=MIN(L1,L2)
+      M=0
+      XM=0.D0
+      SUM=THRJ(XL1,XL2,XL,0.D0,0.D0,0.D0)*PLM(L1,0,CT1)*PLM(L2,0,CT2)
+ 2000 M=M+1
+      IF (M.GT.MTOP) GO TO 3000
+      XM=XM+1.D0
+      SUM=SUM+2.D0*PARITY(M)*THRJ(XL1,XL2,XL,XM,-XM,0.D0)*
+     1             PLM(L1,M,CT1)*PLM(L2,M,CT2)*COS(XM*DP)
+      GO TO 2000
+ 3000 YRR=FACT*SUM
+      RETURN
+ 9999 WRITE(6,699) L1,L2,L
+  699 FORMAT('0 YRR *** ERROR.  ODD ARGUMENTS NOT ALLOWED',3I5)
+      STOP
+      END
+      FUNCTION THRJ(F1,F2,F3,G1,G2,G3)
+      IMPLICIT DOUBLE PRECISION (A-H,O-Z)
+C     SMALL CHANGES 31 JUL 95 (SG)
+      SAVE MUNG,X,Y
+      PARAMETER (MXIX=302)
+      DIMENSION X(MXIX),Y(MXIX)
+      DATA MUNG/0/
+      IF (MUNG.EQ.21) GO TO 69
+      MUNG = 21
+      X(1) = 0.D0
+      DO 100 I = 1, MXIX-1
+      A = I
[…]
