+++ cgqf.f
@@ -1,0 +1,2600 @@
+C            IQPACK - FORTRAN SUBROUTINES FOR THE WEIGHTS OF
+C                       INTERPOLATORY QUADRATURES
+C
+C     FOR A DETAILED DESCRIPTION OF THESE ROUTINES SEE THE PAPER
+C     WITH THE ABOVE TITLE -
+C
+C     GIVEN A SET OF DISTINCT KNOTS, T, AND THEIR MULTIPLICITIES MLT,
+C     THIS PACKAGE COMPUTES THE WEIGHTS D    OF THE INTERPOLATORY
+C                                        J,I
+C     QUADRATURE FORMULA
+C
+C                                            (I)
+C                 SUM       SUM        D    F   (T(J)),
+C               J=1,NT   I=0,MLT(J)-1   J,I
+C
+C            (I)
+C     WHERE F    IS THE I-TH DERIVATIVE OF F, WHERE THE QUADRATURE
+C     IS TO APPROXIMATE
+C
+C                     INTEGRAL  F(T)W(T) DT,
+C                      ÕA,Bå
+C
+C     AND WHERE W(T) IS A WEIGHT FUNCTION. FOR CERTAIN CLASSICAL WEIGHT
+C     FUNCTIONS, LISTED BELOW, NO OTHER INFORMATION IS NEEDED. HOWEVER
+C     THE PACKAGE CAN COMPUTE THE QUADRATURE WEIGHTS CORRESPONDING TO
+C     ANY W(T) FOR WHICH THE ZERO-TH MOMENT AND THE (TRIDIAGONAL
+C     SYMMETRIC) JACOBI MATRIX ASSOCIATED WITH THE POLYNOMIALS
+C     ORTHOGONAL ON ÕA,Bå WITH RESPECT TO W(T), ARE KNOWN. (A UTILITY
+C     ROUTINE IS SUPPLIED TO PROVIDE THIS INFORMATION FOR CLASSICAL
+C     WEIGHT FUNCTIONS). KNOTS AND WEIGHTS OF GAUSS QUADRATURES
+C     WITH NO MULTIPLE KNOTS CAN ALSO BE COMPUTED.
+C
+C     THE PACKAGE IS AN IMPLEMENTATION OF THE METHOD DESCRIBED IN
+C
+C     "CALCULATION OF THE WEIGHTS OF INTERPOLATORY QUADRATURES",
+C     J. KAUTSKY AND S. ELHAY,  NUMER MATH 40 (1982) 407-422,
+C
+C     TOGETHER WITH VARIOUS UTILITY ROUTINES. WEIGHTS TO SOME OR ALL THE
+C     KNOTS CAN BE COMPUTED.
+C
+C           TABLE OF CLASSICAL WEIGHT FUNCTIONS
+C
+C      KIND  INTERVAL       WEIGHT FUNCTION               NAME
+C        1    (A,B)              ONE                    LEGENDRE
+C        2    (A,B)      ((B-X)*(X-A))**(-HALF)         CHEBYSHEV
+C        3    (A,B)      ((B-X)*(X-A))**ALPHA           GEGENBAUER
+C        4    (A,B)    (B-X)**ALPHA*(X-A)**BETA          JACOBI
+C        5   (A,INF)   (X-A)**ALPHA*EXP(-B*(X-A))      GEN LAGUERRE
+C        6  (-INF,INF) ABS(X-A)**ALFA*EXP(-B*(X-A)**2) GEN HERMITE
+C        7    (A,B)      ABS(X-(A+B)/TWO)**ALFA         EXPONENTIAL
[…]
