+++ vecsetuph2co.f
@@ -1,0 +1,960 @@
+c     This file and the accompanying vecstorh2co.f and vecevalh2co.f are a 
+c     replacement for H2-CO potential routine of Jankowski and Szalewicz,
+c     described in (JCP 108, 3554 (1998).
+c     This version must be told in advance the list of angular positions
+c     at which it can expect to be called.  It calculates all
+c     the angle-dependent coefficients that it will need on the 
+c     initial call and stores them in common block /storh2co/.
+c     Then, the potential at the first n of those angular positions
+c     and any distance may be obtained by calling the evaluation
+c     function vecevalh2co() and supplying only the radial distance and
+c     the number of potential values desired.
+
+c     This file contains the initialization code; it was developed
+c     from a routine provided by K. Szalewicz and P. Jankowski.
+c     G. McBane, mcbane@chemistry.ohio-state.edu
+c     29 June 1999
+
+c     In addition to filling the angle arrays
+c     and calling this setup routine, the calling 
+c     program must also insert a value into the variable
+c     Econv (in common block /storh2co/, defined in 
+c     storh2co.f) to determine the units of the returned
+c     energy.  A value of 1.0d0 gives kcal/mol; a value
+c     of 349.755d0 gives cm-1.
+
+      subroutine vecsetuph2co(nangles)
+      implicit double precision (A-H,O-Z)
+      integer nangles
+
+      include 'vecstorh2co.f'
+c      common/exch/lex(3,50),cex(50),nex,irpowex,iwex,iexback,igrx
+
+      if (nangles .gt. maxangles) then
+         write(*, 1000) nangles, maxangles
+         stop
+      end if
+ 1000 format(' Number of angles requested, ', I5, /,
+     *     ' is larger than the maximum number that can be stored, '
+     *     I5, '.',/, 
+     *     '  Increase maxangles in vecstorh2co.f and recompile.')
+
+c     set up for angular calculations
+        call fct(40)
+        call fill3j(13,13,13)
+
+c     store square root factors for calculation of Al1l2l0
+c     this design is an historical remnant from when fill_a_array
+c     was being called for each evaluation
+        call fill_sqrt_array(8)
+
+c     loop over angles, first calculating Al1l2l0 values at
+c     each, then extracting the different coefficients.
+
+        do i = 1, nangles
+           call fill_a_array( cth1vals(i), cth2vals(i), phivals(i) )
+           call fill_asymp( ca(1,i) ) 
+           call fill_short( ga(1,i), expd(i), b15(i) )
+        end do
+
+        return
+        end
+
+c   ------------------------------------------------------
+
+      subroutine fill_short( ga, expd, b15 )
+c  heavily modified from routine potentot() of J&S code
+
+        implicit double precision (a-h,o-z)
+        dimension ga(*)
+        parameter (maxb=400,maxp=3000)
+        parameter (mcsp=maxb)
[…]

[…]
