### 2019-11-27 bc601fa
Add files via upload
added content.tex, metadata.yaml, and bibliography.bib so that formatted article draft can be generated and/or modified by reviewers and editors
