+++ bibliography.bib
@@ -1,0 +1,55 @@
+@STRING{jcp = {J. Chem.\ Phys.}}
+
+@STRING{toms = {ACM Trans.\ Math.\ Soft.}}
+
+
+@ARTICLE{elha87,
+  author = {Sylvan Elhay and Jaroslav Kautsky},
+  title = {Algorithm 655 {IQPACK}: {Fortran} subroutines for the weights of interpolatory
+	quadratures},
+  journal = toms,
+  year = {1987},
+  volume = {13},
+  pages = {399--415},
+  number = {4}
+}
+
+@ARTICLE{gott00,
+  author = {J. Gottfried and George C. McBane},
+  title = {Interaction second virial coefficients from a recent {H$_{2}$--CO}
+	potential energy surface},
+  journal = jcp,
+  year = {2000},
+  volume = {112},
+  pages = {4417--4418},
+  Doi=    {10.1063/1.480989}
+}
+
+@BOOK{hirs54,
+  title = {Molecular Theory of Gases and Liquids},
+  publisher = {Wiley},
+  year = {1954},
+  author = {J. O. Hirschfelder and C. F. Curtiss and R. B. Bird},
+  address = {New York},
+  comment = {({Eq.} (6.5--6) is missing a leading minus sign in the original edition;
+	it is correct in the 1964 printing.)}
+}
+
+@ARTICLE{jank98,
+  author = { Jankowski, P. and Szalewicz, K. },
+  title = { {\textit{Ab initio}} potential energy surface and infrared spectra of {H$_2$--CO} and {D$_2$--CO} van der {Waals} complexes},
+  journal = jcp,
+  year = { 1998 },
+  volume = { 108 },
+  pages = { 3554--3565 }
+}
+
+@ARTICLE{pack83,
+  author = {Russell T. Pack},
+  title = {First quantum corrections to second virial coefficients for anisotropic
+	interactions: Simple, corrected formula},
+  journal = jcp,
+  year = {1983},
+  volume = {78},
+  pages = {7217--7222}
+}
