+++ content.tex
@@ -1,0 +1,36 @@
+\section{Introduction}
+
+For the Reproducibility Challenge I selected a short paper reporting work done in 1999 by Gottfried and McBane \cite{gott00}.  The paper reports the calculation of interaction second virial coefficients between hydrogen and carbon monoxide.  These coefficients are measurable properties of mixtures of the two gases, and reflect the influence of the intermolecular forces between the molecules.  If H$_2$ and CO molecules did not interact, mixtures of them would show exactly the pressure-volume-temperature behavior predicted by the properties of the pure gases and Dalton's Law.  In this case the interaction virial coefficients would be zero at all temperatures. If at a particular temperature the molecules attract one another on average, the pressure of a mixture is  lower than would be predicted on the basis of the pure gas behavior, and the virial coefficient is negative.  If the average intermolecular force is repulsive, the resulting pressure is  above the interpolated pure-gas value, and the virial coefficient is positive.
+
+The interaction virial coefficients can be computed from a model of the intermolecular forces between the two molecules.   Such a model is usually constructed as  a \emph{potential energy surface} (PES), a function describing the difference in energy between pairs of molecules with specified (finite) separation and orientation and the same molecules at infinite separation.  Comparison of computed virial coefficients to measured ones provides a sensitive measure of the quality of a proposed PES.
+
+
+
+The original work reports virial coefficients predicted by the H$_2$--CO PES described by Jankowski and Szalewicz in 1998 \cite{jank98}.  Its Table I shows values at five temperatures where experimental measurements have been reported.  Its  Figure 1 shows a smooth curve constructed from computations at a denser set of temperatures.  Here we concentrate only on the results in the table.
+
+For the case of two rigid diatomic molecules the PES takes the form $V(r,\theta_1, \theta_2, \phi)$, where $r$ is the length of the vector $\mathbf{r}$ between the centers of mass of the two molecules, the two $\theta$ angles describe the orientations of the internuclear axis of each molecule with respect to $\mathbf{r}$, and $\phi$, the dihedral angle, describes the relative rotation of the two internuclear axes about $\mathbf{r}$.  The computation of the virial coefficients from the PES requires a set of multidimensional integrals over the four coordinates $r$, $\theta_1$, $\theta_2$, and $\phi$.  The necessary formulas are given by Pack \cite{pack83} and Hirschfelder, Curtiss, and Bird \cite{hirs54}.  The main contribution to the virial coefficient comes from the ``classical part'' $B_0$.  For  H$_2$--CO, especially at low temperatures, several quantum corrections, requiring different integrals, are also needed.  They make contributions that are not negligible in comparison to experimental uncertainties.  Table I in the original paper reports $B_0$, four different quantum corrections, and their sum, at the five experimental temperatures.
+
+\section{Numerical approach}
+The ``Virial6'' program used for the original work performs the required integrations numerically, using nested Gaussian quadratures.  The angular integrals were done with Gauss-Legendre quadrature.  The integration over $r$ was done in two sections, with an analytic integration from $r=0$ to a small finite distance $r_0$ within the hard core of the potential, and a Gauss-Legendre quadrature in $1/r$ from $r_0$ to infinity.  The Gaussian weights and abscissas were obtained with IQPACK routines \cite{elha87}.  For efficient computation, the integrand values $I_i$ and the composite quadrature weights $w_i$ evaluated at all necessary coordinate sets $\{r_i,\theta_{1i}, \theta_{2i}, \phi_i\}$ were collected into vectors that were principally manipulated using BLAS routines.
+
[…]
