+++ README.md
@@ -5,12 +5,12 @@ This program is intended to reproduce the computed results from Table I of Gottf
 J. Chem. Phys. 112, 4417 (2000).
 
 The source files to be compiled are:
-virial6.f  (main program)
-vecsetuph2co.f and vecevalh2co.f (H2-CO PES setup and evaluation routines, modified from the original potential routines of Jankowski and Szalewicz, JCP 108, 3554 (1998)).
-angular.f  (utility routines evaluating angular functions, taken largely from the MOLSCAT program at https://www.giss.nasa.gov/tools/molscat/)
-cgqf.f (routines from IQPACK for evaluation of Gaussian quadrature abscissas and weights; public source http://www.netlib.org/misc/iqpack)
+1. virial6.f  (main program)
+2. vecsetuph2co.f and vecevalh2co.f (H2-CO PES setup and evaluation routines, modified from the original potential routines of Jankowski and Szalewicz, JCP 108, 3554 (1998)).  Another file, vecstorh2co.f, does not need to be compiled explicitly, but should be present in the directory where the compilation is done.  It is INCLUDEd in the two other vecxx.f files and stores common block information.
 
-A sixth file, vecstorh2co.f, does not need to be compiled explicitly, but should be present in the directory where the compilation is done.  It is INCLUDEd in the code in two other files and stores common block information.
+3. angular.f  (utility routines evaluating angular functions, taken largely from the MOLSCAT program at https://www.giss.nasa.gov/tools/molscat/)
+4. cgqf.f (routines from IQPACK for evaluation of Gaussian quadrature abscissas and weights; public source http://www.netlib.org/misc/iqpack)
+
 
 To compile and link the program, those five Fortran files should be compiled and linked with a double precision BLAS library.  If no local BLAS is available, the reference Fortran BLAS from http://www.netlib.org/blas/ may be used.  Only the double precision real routines are needed.
 
