+++ README.md
@@ -15,7 +15,9 @@ The source files to be compiled are:
 To compile and link the program, those five Fortran files should be compiled and linked with a double precision BLAS library.  If no local BLAS is available, the reference Fortran BLAS from http://www.netlib.org/blas/ may be used.  Only the double precision real routines are needed.
 
 The executable is controlled by an input file read on standard input with READ(*,) statements.  The input file az009f.inp is provided.  If the executable is in the PATH, then on either Linux or Windows (CMD shell, not Powershell) the command
-virial6 <az009f.inp 
+
+> virial6 <az009f.inp 
+
 should produce all the computed values from the published Table.  (Note that the temperatures are in ascending order in the published paper, while they are in descending order in the program output.) The file sample.output shows the desired result.
 
 Questions may be sent to George McBane, mcbaneg at gvsu dot edu, @mcbaneg
