### 2019-11-27 #11 @mcbaneg
Pierre,

There’s now a GitHub repository,
https://github.com/mcbaneg/tenyears-submission,
where you are listed as a collaborator.

Only the master branch exists at the moment.  The code, sample input and output files, and article source information are all there.

Best regards,
George.

George C. McBane (mcbaneg@gvsu.edu<mailto:mcbaneg@gvsu.edu>)
Professor of Chemistry
Assistant Dean for Research, Facilities, and Analytics
College of Liberal Arts and Sciences
Grand Valley State University

B-4-229 Mackinac Hall, (616) 331-2506
Chemistry site http://www.gvsu.edu/chem/
CLAS site http://www.gvsu.edu/clas/
individual site http://faculty.gvsu.edu/mcbaneg/
