### 2019-12-03 #11 @pdebuyl
Thank you @mcbaneg for the update.

I will follow mostly the regular review process that you can read at https://rescience.github.io/write/ and https://rescience.github.io/edit/ , adapting of course to the specifics of the challenge.

I provide a first check list here (the first four items come from ReScience's guidelines).

- [ ] The actual replication of the research. The reviewer must evaluate the authors’ claims about a successful replication, applying the standards of the field.
- [ ] Reproducibility of the replication. The reviewers must be able to run the proposed implementation on their computer, and obtain the same results as the authors with the limits of the state of of the art of the field.
- [ ] Clarity of the code and the instructions for running it. Uncommented or obfuscated code is as bad as no code at all.
- [ ] Clarity and completeness of the accompanying article, in which the authors should clearly state why they think they have replicated the paper (same figures, same graphics, same behavior, etc.) and explain any obstacles they have encountered during the replication work.
- [ ] Past source code: availability and licensing
- [ ] Modifications needed to reproduce the results.
- [ ] Update the source code: availability and licensing.

I'll provide a first report this week so that we can move forward with the review.

Regarding BLAS, I installed it via the Debian package system and added `-lblas` to the gfortran link line :-)
