### 2019-12-03 #11 @mcbaneg
Pierre,

Sounds good.

One of the points on your list seems inconsistent with the Challenge guidelines:

  *   [ ] Clarity of the code and the instructions for running it. Uncommented or obfuscated code is as bad as no code at all.
I would certainly agree with that guideline if the article was publishing new code, even for the purpose of replicating an old study.  But in this Reproducibility Challenge case it seems to me that going back in and cleaning up the old code would constitute cheating: the guidelines explicitly say they want participants to rerun their old programs, not write new code.

I hope you find my code comprehensible and clear and well documented.  I also hope that in the past 20 years I have learned some things about how to write clear and comprehensible programs, so that if I were to redo the project the result would be better than it was the first time.  But for this Challenge I don’t feel like going in and fixing things before submitting would be appropriate. Do you agree?

-G.
