### 2019-12-10 #11 @pdebuyl
@mcbaneg I also believe that the requirements for the ten year challenge should be somehow
relaxed.

Please find below my first report. This is a first case as it is a success and my comments
revolve mostly around our publishing principles: licensing and environment information.

**Replication:** The scope of the results is well presented: the second virial coefficient
for a mixture of two gases. Even though it might seem redundant, I suggest to add the table
with the numerical values. In all ReScience papers, we include the final data in some form,
table or figure.

**Clarity of the code and instructions for running it:** The code is written in a procedural
style, which is expected for Fortran 77. The code is commented sufficiently to follow the
logic.

I could run the code rather easily. I ran `gfortran -c file.f` for all Fortran files except
for the common block file `vecstorh2co.f`, then `gfortran -o virial6 *.o -lblas` and ran the
code as indicated in the README file: `virial6 < az009f.inp`. The instructions should of the
README file are sufficient for Fortran programmers.

gfortran reports a warning when passing a character variable of the inappropriate
length. This only occurs for the label of the error and does not influence any of the
numerical routines.

**Past source code:** The source code is available. Licensing, as far as I could see from
web searches, seems ok for the external code: molscat has been released under the GPL since
the 90's and the IQPACK routines from netlib should be, unless otherwise stated, in the
public domain.

The license for the main program is not defined, this should be fixed.

I believe that you can also comment on the choice of technology, here Fortran. This is
consistent with the aims of the ten year challenge.

An extra request I have concerns the platform definition. Please state, for both the old
environment (if possible) and the new: CPU type, operating system, compiler (including the
version).

There was an example for Python on the previous website
https://github.com/ReScience/rescience.github.io/blob/sources/07-platform-instruction and
some papers provide the information
https://github.com/ReScience-Archives/Senden-Schuecker-Hahne-Diesmann-Goebel-2018/tree/master/code
https://github.com/ReScience-Archives/Detorakis-2017/tree/master/code
