### 2019-12-12 #11 @khinsen
Thanks @pdebuyl! Given your experience with this first review, could you please check if the [author guidelines draft](https://github.com/ReScience/ten-years/blob/master/author-guidelines.md) covers everything you consider important?
