### 2019-12-12 #11 @mcbaneg
@pdebuyl: thank you for the review.  I am working on addressing your comments.

 I have one question: you said
> I believe that you can also comment on the choice of technology, here Fortran. This is
consistent with the aims of the ten year challenge.

I had tried to do that at the end of Section 2, where I wrote

>In molecular physics it was (and is) common practice for groups developing new PESs to provide Fortran subroutines implementing their surfaces as functions of the coordinates.  Such a routine was available for the Jankowski and Szalewicz surface, deposited as supplementary information with their original publication.  A revised version of this routine, more efficient for rapid multiple evaluations, was used for the project.  The  main Virial6 program was written in Fortran for compatibility with such PES routines.  The IQPACK library was then selected as a convenient and well-established Fortran source of accurate numerical quadrature data.

Is that paragraph inadequately clear, or is some other information about the choice of Fortran missing?

Thank you!
G.
