### 2019-12-17 #11 @pdebuyl
@mcbaneg the paragraph quoted informs the reader that Fortran is a common choice in your field of research. There are other criteria for picking a language:
- Are there software resources in that language? You mention IQPACK as an afterthought but the existence of BLAS and of the spherical routines can also be important.
- The ease of implementation, envisioned lasting character and the availability of compilers are other subjective motivations that you can comment on, for instance.

It does not have to be a long explanation and you can pick what you wish as a motivation. As we look at "old" software, I believe that it is important to add context about our technological choices.

Regards,

Pierre
