### 2019-12-20 aea41b1
Update README.md
extended README.md to include compiler and hardware environment for modern reproduction, and what modest information was available about the original (1999) environment
