+++ content.tex
@@ -8,16 +8,16 @@ The interaction virial coefficients can be computed from a model of the intermol
 
 The original work reports virial coefficients predicted by the H$_2$--CO PES described by Jankowski and Szalewicz in 1998 \cite{jank98}.  Its Table I shows values at five temperatures where experimental measurements have been reported.  Its  Figure 1 shows a smooth curve constructed from computations at a denser set of temperatures.  Here we concentrate only on the results in the table.
 
-For the case of two rigid diatomic molecules the PES takes the form $V(r,\theta_1, \theta_2, \phi)$, where $r$ is the length of the vector $\mathbf{r}$ between the centers of mass of the two molecules, the two $\theta$ angles describe the orientations of the internuclear axis of each molecule with respect to $\mathbf{r}$, and $\phi$, the dihedral angle, describes the relative rotation of the two internuclear axes about $\mathbf{r}$.  The computation of the virial coefficients from the PES requires a set of multidimensional integrals over the four coordinates $r$, $\theta_1$, $\theta_2$, and $\phi$.  The necessary formulas are given by Pack \cite{pack83} and Hirschfelder, Curtiss, and Bird \cite{hirs54}.  The main contribution to the virial coefficient comes from the ``classical part'' $B_0$.  For  H$_2$--CO, especially at low temperatures, several quantum corrections, requiring different integrals, are also needed.  They make contributions that are not negligible in comparison to experimental uncertainties.  Table I in the original paper reports $B_0$, four different quantum corrections, and their sum, at the five experimental temperatures.
+For the case of two rigid diatomic molecules the PES takes the form $V(r,\theta_1, \theta_2, \phi)$, where $r$ is the length of the vector $\mathbf{r}$ between the centers of mass of the two molecules, the two $\theta$ angles describe the orientations of the internuclear axis of each molecule with respect to $\mathbf{r}$, and $\phi$, the dihedral angle, describes the relative rotation of the two internuclear axes about $\mathbf{r}$.  The computation of the virial coefficients from the PES requires a set of multidimensional integrals over the four coordinates $r$, $\theta_1$, $\theta_2$, and $\phi$.  The necessary formulas are given by Pack \cite{pack83} and Hirschfelder, Curtiss, and Bird \cite{hirs54}.  The main contribution to the virial coefficient comes from the ``classical part'' $B_0$.  For  H$_2$--CO, especially at low temperatures, several quantum corrections, requiring different integrals, are also needed.  They make contributions that are not negligible in comparison to experimental uncertainties.  Table I in the original paper reports $B_0$, five different quantum corrections, and their sum, at the five experimental temperatures.
 
 \section{Numerical approach}
 The ``Virial6'' program used for the original work performs the required integrations numerically, using nested Gaussian quadratures.  The angular integrals were done with Gauss-Legendre quadrature.  The integration over $r$ was done in two sections, with an analytic integration from $r=0$ to a small finite distance $r_0$ within the hard core of the potential, and a Gauss-Legendre quadrature in $1/r$ from $r_0$ to infinity.  The Gaussian weights and abscissas were obtained with IQPACK routines \cite{elha87}.  For efficient computation, the integrand values $I_i$ and the composite quadrature weights $w_i$ evaluated at all necessary coordinate sets $\{r_i,\theta_{1i}, \theta_{2i}, \phi_i\}$ were collected into vectors that were principally manipulated using BLAS routines.
 
[…]
