### 2019-12-20 #11 @mcbaneg
@pdebuyl and @khinsen: I have modified the submission in response to the comments:

> Even though it might seem redundant, I suggest to add the table
with the numerical values. In all ReScience papers, we include the final data in some form,
table or figure.

Done, now Table 1 in section 4.

> Licensing, as far as I could see from
web searches, seems ok for the external code: molscat has been released under the GPL since
the 90's and the IQPACK routines from netlib should be, unless otherwise stated, in the
public domain.
The license for the main program is not defined, this should be fixed.

I have added a LICENSE.txt file.  I did not want to assign new licenses to code produced by others so I have linked to the relevant licenses as best I could.  The LICENSE.txt also includes a note explaining that the CO_H2 potential surface is the work of Jankowski and Szalewicz, and use of it in published work requires appropriate citation independent of the license that applies to the software used to implement it.

> I believe that you can also comment on the choice of technology, here Fortran. This is
consistent with the aims of the ten year challenge.

I have added a modest amount to the material at the end of section 2 in response.

> An extra request I have concerns the platform definition. Please state, for both the old
environment (if possible) and the new: CPU type, operating system, compiler (including the
version).

I have added this information to the README.md file.
