### 2019-12-20 #11 @pdebuyl
@mcbaneg Thank you for the update, I am happy with the modifications.

Before finalizing the review, I only have two remarks:
1. The comment about mandatory citation of the 1998 paper is, as far as I can tell, not compatible with the MIT license.
2. You can still decide to archive your code either on zenodo or on software heritage.

@khinsen @rougier can you comment on the license issue here? Also, once the article is accepted, do you take the process over or do I need to do something?
