### 2019-12-20 #11 @mcbaneg
@khinsen,  @pdebuyl, @rougier ,

> The comment about mandatory citation of the 1998 paper is, as far as I can tell, not compatible with the MIT license.

I'd like advice on how to do this correctly.  I want to distinguish between the code I wrote and the original H2-CO potential surface of Jankowski and Szalewicz that they constructed and described in their paper.  Ordinary scientific ethics requires that anyone who uses the J&S surface describing H2-CO and publishes results built with it must acknowledge J&S with a citation.  This is true whether they use the routine that J&S originally supplied, my program derived from it, or any other evaluation mechanism including a new one written for the purpose.

My code provides a particular implementation of their surface. A permissive license applied to my code does not relieve the user of the obligation to acknowledge J&S in publications describing use of their work.  I was trying to convey that idea with the statement

> The files vectorh2co.f, vecsetuph2co.f, and vecevalh2co.f implement the H2-CO potential energy surface of P. Jankowski and K. Szalewicz, Journal of Chemical Physics 108, 3554 (1998).  Any published work that uses their surface must cite that paper.

at the top of LICENSE.txt.

How do I do this right?

Thank you!
