### 2019-12-21 #11 @khinsen
My point of view on the licensing question is that licenses and citation requirement live in different universes. LIcenses are legally binding and apply to software as a tool. Citation requirements are moral (I haven't heard of any case going to court) and apply more to methods than to implementations. The idea of citing software itself is a recent invention and (as I see it!) just a workaround to make a broken system of credit a bit less negative for software developers.

I agree with @pdebuyl that a requirement to cite anything is incompatible with the MIT license, for two reasons: 1) the MIT license has no provision for such restrictions, 2) the MIT license is about *using* and *copying* software, whereas a citation request is about *publishing* results obtained with the software.

So I'd keep citation out of the license but add a separate CITATION file (see https://www.software.ac.uk/blog/2017-12-12-standard-format-citation-files) to establish the moral requirements for academic use.
