### 2020-01-26 fc1c0cd
Add files via upload
README converted to .md format, note about appropriate citation of J&S surface added
CITATION.cff modified to described only software citation
LICENSE.txt modified to remove comment about citation of J&S
