+++ README.md
@@ -1,41 +1,50 @@
-# tenyears-submission
-McBane's Ten Years Reproducibility Challenge submission
+## Virial6
+As provided, this program is intended to reproduce the computed results from Table I of Gottfried and McBane, J. Chem. Phys. 112, 4417 (2000). These report interaction second virial coefficients for hydrogen-carbon monoxide mixtures, using the potential surface of Jankowski and Szalewicz, J. Chem. Phys. 108, 3554 (1998).  Virial6 can be used to compute second virial coefficients for other diatom-diatom (or smaller) systems if the potential energy surface evaluation routines are modified appropriately.  
 
-This program is intended to reproduce the computed results from Table I of Gottfried and McBane,
-J. Chem. Phys. 112, 4417 (2000).
+Any published work that uses the Jankowski and Szalewicz potential surface (as implemented by the routines provided here or others) must cite their paper.
 
+### Compilation and linking
 The source files to be compiled are:
-1. virial6.f  (main program)
-2. vecsetuph2co.f and vecevalh2co.f (H2-CO PES setup and evaluation routines, modified from the original potential routines of Jankowski and Szalewicz, JCP 108, 3554 (1998)).  Another file, vecstorh2co.f, does not need to be compiled explicitly, but should be present in the directory where the compilation is done.  It is INCLUDEd in the two other vecxx.f files and stores common block information.
+- virial6.f  (main program)
+- vecsetuph2co.f and vecevalh2co.f (H2-CO PES setup and evaluation routines, modified from the original potential routines of Jankowski and Szalewicz, J. Chem. Phys. 108, 3554 (1998)). An additional file, vecstorh2co.f, does not need to be compiled explicitly, but should be present in the directory where the compilation is done.  It is INCLUDEd in the code in two other files and stores common block information.  
+- angular.f  (utility routines evaluating angular functions, taken largely from the MOLSCAT program at https://www.giss.nasa.gov/tools/molscat/)
+- cgqf.f (routines from IQPACK for evaluation of Gaussian quadrature abscissas and weights; public source http://www.netlib.org/misc/iqpack)
 
-3. angular.f  (utility routines evaluating angular functions, taken largely from the MOLSCAT program at https://www.giss.nasa.gov/tools/molscat/)
-4. cgqf.f (routines from IQPACK for evaluation of Gaussian quadrature abscissas and weights; public source http://www.netlib.org/misc/iqpack)
 
 
 To compile and link the program, those five Fortran files should be compiled and linked with a double precision BLAS library.  If no local BLAS is available, the reference Fortran BLAS from http://www.netlib.org/blas/ may be used.  Only the double precision real routines are needed.
 
-The executable is controlled by an input file read on standard input with READ(*,) statements.  The input file az009f.inp is provided.  If the executable is in the PATH, then on either Linux or Windows (CMD shell, not Powershell) the command
+### Use
 
-> virial6 <az009f.inp 
-
+The executable is controlled by an input file read on standard input with READ(*,) statements.  The input file az009f.inp is provided.  If the executable is in the PATH, then on either Linux or Windows (CMD shell, not Powershell) the command  
+virial6 <az009f.inp  
[…]
