+++ README.md
@@ -26,15 +26,15 @@ The program is written in near-standard Fortran 77 with a few more modern extens
 
 The 1999 work used the g77 compiler (probably version 0.5.2x, based on gcc 2.95) with the reference Fortran BLAS on 32-bit Microsoft Windows (probably Windows 95). Details are no longer easily accessible.
 
-For the 2019 reproduction, the environment was
-Compiler: Intel Visual Fortran
-Compiler version: 19.0.3.203
-Compiler flags:  -O3  -traceback
+For the 2019 reproduction, the environment was  
+Compiler: Intel Visual Fortran  
+Compiler version: 19.0.3.203  
+Compiler flags:  -O3  -traceback  
 OS: 64-bit Microsoft Windows 10 Enterprise version 1903
 
-Hardware:
-model name      : Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
-vendor_id       : GenuineIntel
+Hardware:  
+model name      : Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz  
+vendor_id       : GenuineIntel  
 flags           : fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush dts acpi mmx fxsr sse sse2 ss ht tm pbe pni dtes64 est tm2 ssse3 fma cx16 xtpr pdcm sse4_1 sse4_2 movbe popcnt aes xsave osxsave avx f16c rdrand hypervisor lahf_lm ida epb xsaveopt pln pts dtherm fsgsbase tsc_adjust bmi1 hle avx2 smep bmi2 erms invpcid rtm mpx rdseed adx smap clflushopt
 
 
