### 2020-01-26 #11 @mcbaneg
@khinsen, @pdebuyl, @rougier ,

I have now modified my submission to try to address all the comments.  In addition to the changes described in https://github.com/ReScience/submissions/issues/11#issuecomment-567787725,

1. I have removed comments about citing Jankowski and Szalewicz from the LICENSE file, and added them as a note in the README.md.  My study of the CITATION.cff format indicated that it was specifically aimed at software/code citations, not to the ordinary publication/citation mechanism.
2. I did add a CITATION.cff file applying to Virial6 itself.
3.  I added a PLATFORM information section to README.md.
4.  I have tried to extend the description of the choice of Fortran for the project.

I have not submitted the code explicitly to either Software Heritage or Zenodo.  I did not want to duplicate the automatic acquisition of the GitHub project by Software Heritage, or any further processes done by ReScience C.  Once again I will welcome recommendations about what action might be best.

Thank you!
