### 2020-01-27 #11 @khinsen
Submitting your GitHub repository to Software Heritage will only make it archived earlier, so it can do no harm. Software Heritage does automatic deduplication in the same way that git does: via content-addressable storage. In fact, we added the recommendation to submit to Software Heritage at the suggestion of the Software Heritage team!
