### 2020-01-28 #11 @pdebuyl
@mcbaneg @khinsen  @rougier

I have no further comments or questions on the manuscript and I recommend publication.

@khinsen  @rougier What's next in the process?
