### 2020-01-28 #11 @rougier
@khinsen will need now to actually publish the paper which is not totally straightforward yet. @mcbaneg will need to update the metadata with the information provided by @khinsen and we may need to transform the template a bit to have software heritage identifier on code (instead of DOI).
