### 2020-01-28 #11 @khinsen
I'll try to do all of that tomorrow. Thanks @rougier for pointing out the SWH identifier, that's indeed a new feature but one that should also be useful for future regular articles.
