### 2020-01-29 7711573
Change article type from "replication" to "reproduction"
