+++ metadata.yaml
@@ -5,7 +5,7 @@
 #  - For a successful replication, it shoudl be prefixed with "[Re]"
 #  - For a failed replication, it should be prefixed with "[¬Re]"
 #  - For other article types, no instruction (but please, not too long)
-title: "[Re] Reproduction of interaction second virial coefficient calculation for H$_2$--CO interactions [J. Chem. Phys. vol. 112, 4417 (2000)]"
+title: "[Rp] Reproduction of interaction second virial coefficient calculation for H$_2$--CO interactions [J. Chem. Phys. vol. 112, 4417 (2000)]"
 
 # List of authors with name, orcid number, email and affiliation
 # Affiliation "*" means contact author
@@ -60,7 +60,7 @@ bibliography: bibliography.bib
 #  * Editorial
 #  * Letter
 #  * Replication
-type: Replication
+type: Reproduction
 
 # Scientific domain of the article (e.g. Computational Neuroscience)
 #  (one domain only & try to be not overly specific)
