### 2020-01-29 #11 @khinsen
The article is published: https://zenodo.org/record/3630224

Congratulations to @mcbaneg and thanks to @pdebuyl for the review!

@mcbaneg: Please accept pull request https://github.com/mcbaneg/tenyears-submission/pull/1 to your code/article repository to update the metadata and the final article PDF.
